spikes/vcfmongo/src/requirements.txt
pymongo==3.2.2
PyVCF==0.6.8
